Move call_api into trackman.view_utils
